requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.0.9
django-cors-headers==4.4.0
django-storages==1.14.4
djangorestframework==3.15.2
idna==3.10
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.3